% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comma}
\alias{comma}
\title{function to format numbers to text with 1,000s comma}
\usage{
comma(number)
}
\arguments{
\item{number}{value to format}
}
\description{
function to format numbers to text with 1,000s comma
}
