% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_table.R
\name{main_table}
\alias{main_table}
\title{create the main safe table}
\usage{
main_table(data, year, tier, c1, c2, c3)
}
\arguments{
\item{data}{a specific dataframe holding results}

\item{year}{what is the current year}

\item{tier}{what tier}

\item{c1}{estimated catch current year}

\item{c2}{proj catch year + 1}

\item{c3}{proj catch year + 2}
}
\description{
create the main safe table
}
