\name{p53.stat}
\docType{data}
\alias{p53.stat}
\title{p53 Mutation Status}
\description{
This data set is included for use in the vignette and provides the p53 mutation status (p53+ = 1 and p53- = 0) for each of 251 samples in the Miller et al. breast cancer study data.
}
\usage{data(p53.stat)}
\format{A data.frame with 2 columns, "samplename" and "p53", and 251 rows.}
\source{NCBI's Gene Expression Omnibus, accession GSE3494}
\references{
Miller, L.D., Smeds, J., George, J., Vega, V.B., Vergara, L., Ploner, A., Pawitan, Y.,Hall, P., Klaar, S., Liu, E.T., and Bergh, J. (2005) `An expression signature for p53status in human breast cancer predicts mutation status, transcriptional effects, and patient survival', \emph{Proc Natl Acad Sci U S A}, 102(38), 13550-13555.
}
\keyword{datasets}