\name{safe-internal}
\alias{local.t.Welch}
\alias{local.t.Student}
\alias{local.t.paired}
\alias{local.t.LM}
\alias{local.f.ANOVA}
\alias{global.Wilcoxon}
\alias{global.Kolmogorov}
\alias{global.AveDiff}
\alias{global.Fisher}
\alias{global.Pearson}
\alias{error.FDR.YB}
\alias{error.FDR.BH}
\alias{error.FWER.WY}
\alias{error.FWER.Bonf}
\alias{error.FWER.Holm}
\alias{getPImatrix}
\alias{getPIcomplete}
\alias{getCOXresiduals}
\alias{getXYresiduals}
\alias{sigfig}

\title{ Internal SAFE functions }
\description{
  The following functions return the appropriate functions within a SAFE analysis and are intended for internal use only.
}
\author{William T. Barry}
\details{
  These are not to be called by the user.
}
\keyword{internal}