\name{safe-package}
\Rdversion{1.1}
\alias{safe-package}
\docType{package}
\title{
Significance Analysis of Function and Expression
}
\description{
SAFE is a resampling-based method for testing functional categories in gene expression experiments. SAFE can be applied to 2-sample and multi-class comparisons, or simple linear regressions. Other experimental designs can also be accommodated through user-defined functions.
}
\details{
\tabular{ll}{
Package: \tab safe\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 2012-09-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
For an overview of how to use the package, including the most important functions, please see the included vignette.
}
\author{
William T. Barry

Maintainer: William T. Barry <bbarry@jimmy.harvard.edu>
}
\references{
W. T. Barry, A. B. Nobel and F.A. Wright, 2005, \emph{Significance Analysis of functional categories in gene expression studies: a structured permutation approach}, \emph{Bioinformatics} \bold{21}(9) 1943--1949. 

}
\keyword{ package }
\seealso{
\code{\link{safe}}, 
}
