\name{safe-internal}
\alias{local.t.Welch}
\alias{local.t.Student}
\alias{local.t.SAM}
\alias{local.t.LM}
\alias{local.f.ANOVA}
\alias{global.Wilcoxon}
\alias{global.Kolmogorov}
\alias{global.genelist}
\alias{error.FDR.YB}
\alias{error.FWER.WY}

\title{ Internal SAFE functions }
\description{
  The following functions return the appropriate functions within a SAFE analysis 
  and are intended for internal use only.
}
\author{William T. Barry}
\details{
  These are not to be called by the user.
}
\keyword{internal}

